Added database types, added script for type conversions
