--- leetcode-tracker/app/_types/commonTypes.ts
@@ -1,11 +1,0 @@
-
-
-export type ProblemType = {
-    id: number;
-    title: string;
-    link: string;
-    difficulty: string;
-    topic: string;
-    goalTime: number;
-    bestTime: number;
-  }
